[LILAC - TEXTURE BUILDER]
Removed the DirectXTex dependency for the texture builder. This will be optional. For now it uses STB by default. Currently does not do any compression.
Changed the texture builder so it now only loads in the header during initialization. It then gets the actual data when it first uses the resource.